+++ lib/community/widgets/post_card.dart
@@ -239,6 +239,7 @@ class _PostCardState extends State<PostCard> {
                 actionsToInclude: [
                   PostCardAction.visitInstance,
                   PostCardAction.visitProfile,
+                  PostCardAction.blockUser,
                   PostCardAction.blockInstance,
                   PostCardAction.visitCommunity,
                   PostCardAction.blockCommunity,